Update StudentTaskSubmission.java: nota aceita null

+++ backend/src/main/java/com/metis/backend/subjects/model/entity/StudentTaskSubmission.java
@@ -11,7 +11,7 @@ public class StudentTaskSubmission {
 
     private String studentUserId;
 
-    private double score;
+    private Float score;
 
     private LocalDateTime submittedIn;
